2nd_week/task4/pyproject.toml
[tool.poetry]
name = "task4"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.34.108"
python-dotenv = "^1.0.1"
python-magic = "^0.4.27"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
